server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==8.3.1
pytesseract==0.3.8
wikipedia==1.4.0
requests==2.26.0
SpeechRecognition==3.8.1
pyttsx3==2.90